Extract PLY file export and byte-size formatting
engine/src/plyExport.ts: savePlyFile and generatePlyFileContent -
rebuilds a PLY file (header, vertex/color/normal data, faces) from the
current transformed geometry, respecting the Uint8-vs-Float32 color
storage difference between point clouds and meshes.

engine/src/utils/format.ts: formatFileSize, already fully pure.

main.ts: 10,988 -> 10,860 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
