+++ engine/src/utils/format.ts
@@ -1,0 +1,13 @@
+export function formatFileSize(bytes: number | undefined): string {
+  if (!bytes) {
+    return 'Unknown';
+  }
+  const units = ['B', 'KB', 'MB', 'GB'];
+  let size = bytes;
+  let unitIndex = 0;
+  while (size >= 1024 && unitIndex < units.length - 1) {
+    size /= 1024;
+    unitIndex++;
+  }
+  return `${size.toFixed(2)} ${units[unitIndex]}`;
+}
